Server/requirements.txt
fastapi
uvicorn
pytesseract
Pillow
huggingface_hub
python-multipart
